Refactor weather system, UBO management, and attribute-less meshes
This commit completes the overhaul of the weather system and its
integration with the engine's resource management patterns.

Key changes:
- Created `WeatherSystem` to manage weather state and transitions,
  decoupling it from rendering logic.
- Refactored `WeatherRenderer` into a thin facade that coordinates
  UBO updates and mesh rendering.
- Introduced `UboManager` to centralize UBO lifecycle management,
  supporting lazy regeneration via registered rebuild functions.
- Extracted `FullScreenMesh` template into `AttributeLessMeshes.h` to
  standardize rendering for shaders that use `gl_VertexID` instead of
  vertex attributes.
- Simplified noise texture initialization by removing unnecessary
  mipmaps for fuzzy atmosphere effects.
- Unified animation timing with the global `AnimationManager`, fixing
  framerated-dependent simulation bugs.
- Resolved various compiler warnings, including weak-vtables and
  formatting issues.
