+++ src/opengl/legacy/AttributeLessMeshes.cpp
@@ -1,0 +1,12 @@
+// SPDX-License-Identifier: GPL-2.0-or-later
+// Copyright (C) 2026 The MMapper Authors
+
+#include "AttributeLessMeshes.h"
+
+namespace Legacy {
+
+BlitMesh::~BlitMesh() = default;
+
+PlainFullScreenMesh::~PlainFullScreenMesh() = default;
+
+} // namespace Legacy
